move eventRegistrations out of account & stored payments + modify UX for OTP modal
