+++ cartridges/int_bolt_embedded_sfra/cartridge/client/default/js/boltStoredPayments.js
@@ -1,43 +1,5 @@
 'use strict';
 
-// register the event listener to the radio buttons for selecting stored payments
-$(document).ready(function () {
-    const radioButtonLoaded = setInterval(function () {
-        const useExistingCardBtn = $('#use-existing-card-radio-button');
-        const addNewCardBtn = $('#add-a-new-card-radio-button');
-        if (useExistingCardBtn && addNewCardBtn) {
-            clearInterval(radioButtonLoaded);
-            useExistingCardBtn.click(function () {
-                if (this.checked) {
-                    $('#bolt-stored-payment-selector').removeClass('d-none');
-                    $('.bolt-pay').addClass('d-none');
-                    setBoltStoredPaymentMethodID();
-                }
-            });
-
-            addNewCardBtn.click(function () {
-                if (this.checked) {
-                    $('#bolt-stored-payment-selector').addClass('d-none');
-                    $('.bolt-pay').removeClass('d-none');
-                    $('#bolt-selected-payment-id').val('');
-                }
-            });
-        }
-    }, 100);
-
-    const boltPaySelectorLoaded = setInterval(function () {
-        const boltPaySelector = $('#bolt-stored-payment-selector');
-        if (boltPaySelector) {
-            clearInterval(boltPaySelectorLoaded);
-            setBoltStoredPaymentMethodID();
-            boltPaySelector.change(function () {
-                setBoltStoredPaymentMethodID();
-                updateBillingAddress();
-            });
-        }
-    }, 100);
-});
-
 /**
  * Adding bolt pay id to payment content so that it's sent to auth when clicking "Place Order"
  * @returns {void}
@@ -104,3 +66,8 @@ function getSelectedCardBillingAddress() {
     }
     return null;
 }
+
+module.exports = {
+    setBoltStoredPaymentMethodID: setBoltStoredPaymentMethodID,
+    updateBillingAddress: updateBillingAddress
+};
