+++ cartridges/int_bolt_embedded_sfra/cartridge/client/default/js/util.js
@@ -9,6 +9,17 @@ function getISOCodeByLocale(locale) {
     return locale.replace('_', '-').toLowerCase();
 }
 
+/**
+ * Checks if the email value entered is correct format
+ * @param {string} email - email string to check if valid
+ * @returns {boolean} Whether email is valid
+ */
+function validateEmail(email) {
+    var regex = /^[\w.%+-]+@[\w.-]+\.[\w]{2,6}$/;
+    return regex.test(email);
+}
+
 module.exports = {
-    getISOCodeByLocale: getISOCodeByLocale
+    getISOCodeByLocale: getISOCodeByLocale,
+    validateEmail: validateEmail
 };
